// features/auth/schemas/register.schema.ts
import { z } from "zod";

export const registerSchema = z.object({
  name: z.string().min(3, "Name must be at least 3 characters"),

  email: z.email("Invalid email"),

  password: z.string().min(6, "Password must be at least 6 characters"),

  role: z.enum(["admin", "user"]),
});

export type RegisterSchema = z.infer<typeof registerSchema>;